+++ Y2F/@optimizerFORCES/compileSolverInterfaceCode.m
@@ -119,7 +119,7 @@ if exist( [cName '.c'], 'file' ) && exist( [mexName '.c'], 'file' )
         
         % figure our whether we need additional libraries for Intel
         clientPath = fileparts(which('generateCode'));
-        intelLibsDir = [clientPath,filesep,'libs_intel'];
+        intelLibsDir = [clientPath,filesep,'libs_Intel'];
         if( exist( intelLibsDir, 'dir' ) && ~isMinGW)
             % If subdirectory for specific architecture exisst, we need to
             % use that
@@ -128,7 +128,6 @@ if exist( [cName '.c'], 'file' ) && exist( [mexName '.c'], 'file' )
                 intelLibsDir = archDir;
             end
             intelLibsDirFlag = ['-L', intelLibsDir];
-            addpath(intelLibsDir); savepath;
         else
             intelLibsDirFlag = '';
         end